Fix the problem with early isContractDeployed checking in step9
As this tutorial is going to work for the folks that have just created their wallets, obviously there is not any contract deployed to their wallet yet. So checking `if (!await client.isContractDeployed(wallet.address))` is a never-happening statement and no one can go further with this tutorial except if they have a wallet that had previous OUT transfers.